REFACTOR: Add utils module
