+++ main.py
@@ -7,58 +7,15 @@ argument 3: The end date to retrieve data (e.g '2022-05-27')
 """
 
 import argparse
-import csv
 import json
-from datetime import datetime
-import time
 import requests
-import re
 import pandas as pd
-import sys
-import getopt
+from libs.utils import HumanDateToEpoch, crypto_asset_json_to_list
 
 # TODO - Later on, add table for currencies and there value in the main coins (USD, Euro and Sterling)
 # TODO - Do same process of cryptos for stocks
 # TODO - Create subdirectories to better organize project
 # TODO - Create tests for project
-
-
-class HumanDateToEpoch:
-
-    # Put in utils module
-
-    @staticmethod
-    def __verify_format(human_date_string: str) -> bool:
-
-        date_pattern = "[0-9]{4}-[0-9]{2}-[0-9]{2}"
-
-        return True if re.match(date_pattern, human_date_string) else False
-
-    @staticmethod
-    def __is_future_date(human_date: str) -> bool:
-
-        utc_date = datetime.utcnow().date()
-
-        return False if utc_date < human_date else True
-
-    @staticmethod
-    def date_to_unix_miliseconds(human_date_string: str) -> int:
-
-        try:
-            dt_date = datetime.strptime(human_date_string, "%Y-%m-%d").date()
-        except ValueError:
-            print("ERROR: Inserted date is not valid. Must be in format YYYY-mm-dd.\n")
-            exit()
-
-        if not HumanDateToEpoch.__is_future_date(dt_date):
-            print("ERROR: Input date is in the future. Correct and try again.\n")
-            exit()
-
-        result_epoch_string = str(
-            int(time.mktime(dt_date.timetuple()) * 1000)  # int to remove float decimal
-        )
-
-        return result_epoch_string
 
 
 def get_crypto_asset_history(asset_id: str, start_date: str, end_date: str) -> dict:
@@ -72,24 +29,6 @@ def get_crypto_asset_history(asset_id: str, start_date: str, end_date: str) -> d
     response = requests.request("GET", crypto_asset_api_url)
 
     return json.loads(response.text.encode("utf8"))
-
-
-def crypto_asset_json_to_list(crypto_request_json: dict) -> list:
-
-    # put in utils module
-
-    try:
-        crypto_asset_data_dict = crypto_request_json["data"]
-    except KeyError:
-        print("ERROR-> " + crypto_request_json["error"])
-        sys.exit(2)
-
-    list_crypto_dicts = []
-
-    for data in crypto_asset_data_dict:
-        list_crypto_dicts.append(data)
-
-    return list_crypto_dicts
 
 
 def process_crypto_dataframe(asset_id: str, list_crypto_dicts: list) -> pd.DataFrame:
